fix(api-reference): ensure section is open and operation has loaded before scrolling to it
