+++ packages/api-reference/src/components/Content/Tag/Endpoints.vue
@@ -16,21 +16,13 @@ import {
 
 const props = defineProps<{ id?: string; tag: Tag }>()
 
-const emit = defineEmits<{
-  (event: 'observeAndNavigate', value: string): void
-}>()
-
 const { getOperationId, getTagId } = useNavState()
-const { setCollapsedSidebarItem } = useSidebar()
-
-const navigateToOperation = (operationId: string) => {
-  emit('observeAndNavigate', operationId)
-}
+const { scrollToIdWhenLoaded } = useSidebar()
 
 // TODO in V2 we need to do the same loading trick as the initial load
-async function scrollHandler(operation: TransformedOperation) {
-  navigateToOperation(getOperationId(operation, props.tag))
-  setCollapsedSidebarItem(getTagId(props.tag), true)
+const scrollHandler = async (operation: TransformedOperation) => {
+  const id = getOperationId(operation, props.tag)
+  scrollToIdWhenLoaded(id)
 }
 </script>
 <template>
