--- packages/api-reference/src/components/Content/Tag/mutationObserver.ts
@@ -1,9 +1,0 @@
-import { useMutationObserver } from '@vueuse/core'
-import type { Ref } from 'vue'
-
-export const observeMutations = (
-  target: Ref<HTMLElement | null>,
-  callback: MutationCallback,
-) => {
-  useMutationObserver(target, callback, { childList: true, subtree: true })
-}
